+++ docs/changelog.md
@@ -16,6 +16,7 @@ Click on the version in the following list to see all changes since the previous
 * chore: switch to Xenial distribution for Travis CI build
 * chore: fix the spelling of Scrum Poker in comments and changes
 * test: remove todo for deployment with licensing disabled after successful test
+* feat: health check for Scrum Poker app available in configuration (#67)
 
 ### [4.0.1] - 2019-06-16